+++ packages/ripple-tide-api/types.d.ts
@@ -1,4 +1,5 @@
 import type { AxiosInstance } from 'axios'
+import { TideAlert } from './src/mapping/alerts/site-alerts-mapping'
 import { TideContact } from './src/mapping/sidebar-contacts/sidebar-contacts-mapping-types'
 import { TideTopicTag } from './src/mapping/topic-tags/topic-tags-mapping'
 
@@ -8,13 +9,52 @@ export interface RplTideModuleMappingFunction {
   // eslint-disable-next-line @typescript-eslint/ban-types
   [key: string]: Function | string | string[] | object
 }
-
 export interface RplTideMapping {
   component?: string | string[]
   schema?: string
   key?: string
   mapping: RplTideModuleMappingFunction
   includes: string[]
+}
+
+export interface TideSiteData {
+  name: string
+  _src?: any
+  siteAlerts: TideAlert[]
+  siteLogo: {
+    href: string
+    src: string
+    altText: string
+  }
+  showQuickExit: boolean
+  acknowledgementHeader?: string
+  acknowledgementFooter: string
+  copyrightHtml: string
+  footerLogos: {
+    alt: string
+    url: string
+    src: string
+  }[]
+  theme: {
+    [key: string]: string
+  }
+  socialImages: {
+    twitter: any
+    og: any
+  }
+  menus: {
+    menuMain: TideMenuItem[]
+    menuFooter: TideMenuItem[]
+  }
+}
+
+export interface TideMenuItem {
+  text: string
+  url: string
+  id: string
+  parent: string | null
+  weight: number
+  items: TideMenuItem[]
 }
 
 export interface TideUrlField {