+++ packages/ripple-tide-media/src/pages/audio/tide-media-audio.vue
@@ -37,12 +37,13 @@ export default { name: 'TideMediaAudioPage' }
 </template>
 
 <script setup lang="ts">
+import { TideSiteData } from '@dpc-sdp/ripple-tide-api/types'
 import type { TideMediaPage } from '../../types'
 import TideMediaHeader from '../../components/tide-media-header.vue'
 import TideMediaBody from '../../components/tide-media-body.vue'
 
 interface Props {
-  site: any
+  site: TideSiteData
   page: TideMediaPage
 }