+++ packages/ripple-tide-news/src/components/tide-news-header.vue
@@ -7,7 +7,7 @@ export default { name: 'TideNewsHeader' }
     :corner-top="true"
     :corner-bottom="true"
     :behind-nav="true"
-    :breadcrumbs="true"
+    :breadcrumbs="hasBreadcrumbs"
     :title="header.title"
   >
     <p class="rpl-type-p-large">{{ header.summary }}</p>
@@ -20,5 +20,6 @@ import type { TideNewsHeader } from '../types'
 
 defineProps<{
   header: TideNewsHeader
+  hasBreadcrumbs: boolean
 }>()
 </script>